+++ t/snippets30.t
@@ -17,6 +17,8 @@
 #14 git159.git159
 #15 git162.def
 #16 git162.git162
+#17 qwaf.def
+#18 qwaf.qwaf
 
 # To locate test #13 you can search for its name or the string '#13'
 
@@ -49,6 +51,11 @@ BEGIN {
 ----------
         'git159' => "-bl -nsbl",
         'git162' => "-nwrs=A",
+        'qwaf'   => <<'----------',
+# git164
+-qwaf
+-sfp
+----------
     };
 
     ############################
@@ -142,6 +149,36 @@ match( $n : == ) {
 
         'logical_xor' => <<'----------',
 $x^^$y and say "One of x or y is true, but not both";
+----------
+
+        'qwaf' => <<'----------',
+use Digest::MD5 qw( md5_hex );
+
+@fields = qw( $st_dev	   $st_ino    $st_mode
+      $st_nlink   $st_uid    $st_gid
+      $st_rdev    $st_size
+      $st_atime   $st_mtime  $st_ctime
+      $st_blksize $st_blocks
+);
+
+@hdr_colors = qw(
+        CadetBlue1
+        MediumPurple1
+        turquoise1
+        PaleTurquoise1
+        SlateBlue1
+    );
+
+# has blank line, so keep line breaks
+@hdr_colors = qw(
+
+        CadetBlue1
+        MediumPurple1
+        turquoise1
+        PaleTurquoise1
+        SlateBlue1
+    );
+
 ----------
     };
 
@@ -383,6 +420,68 @@ match( $n :== ) {
     default             { say "It's something else" }
 }
 #16...........
+        },
+
+        'qwaf.def' => {
+            source => "qwaf",
+            params => "def",
+            expect => <<'#17...........',
+use Digest::MD5 qw( md5_hex );
+
+@fields = qw( $st_dev	   $st_ino    $st_mode
+  $st_nlink   $st_uid    $st_gid
+  $st_rdev    $st_size
+  $st_atime   $st_mtime  $st_ctime
+  $st_blksize $st_blocks
+);
+
+@hdr_colors = qw(
+  CadetBlue1
+  MediumPurple1
+  turquoise1
+  PaleTurquoise1
+  SlateBlue1
+);
+
+# has blank line, so keep line breaks
+@hdr_colors = qw(
+
+  CadetBlue1
+  MediumPurple1
+  turquoise1
+  PaleTurquoise1
+  SlateBlue1
+);
+
+#17...........
+        },
+
+        'qwaf.qwaf' => {
+            source => "qwaf",
+            params => "qwaf",
+            expect => <<'#18...........',
+use Digest::MD5 qw(md5_hex);
+
+@fields = qw(
+    $st_dev  $st_ino  $st_mode  $st_nlink $st_uid   $st_gid
+    $st_rdev $st_size $st_atime $st_mtime $st_ctime $st_blksize
+    $st_blocks
+);
+
+@hdr_colors =
+  qw( CadetBlue1 MediumPurple1 turquoise1 PaleTurquoise1 SlateBlue1 );
+
+# has blank line, so keep line breaks
+@hdr_colors = qw(
+
+    CadetBlue1
+    MediumPurple1
+    turquoise1
+    PaleTurquoise1
+    SlateBlue1
+);
+
+#18...........
         },
     };